backend/requirements.txt
# MedAI Detect — Backend Requirements
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
torch==2.4.1
torchvision==0.19.1
Pillow==10.4.0
numpy==1.26.4
scikit-learn==1.5.2
opencv-python-headless==4.10.0.84
httpx==0.27.2
pydantic==2.9.2
matplotlib==3.9.2
python-dotenv==1.0.1
scipy==1.14.0
tqdm==4.66.5
sqlalchemy==2.0.23
